+++ packages/wow-icons/src/component/index.ts
@@ -1,7 +1,14 @@
+export { default as Calendar } from "./Calendar.tsx";
 export { default as Check } from "./Check.tsx";
 export { default as Close } from "./Close.tsx";
 export { default as DownArrow } from "./DownArrow.tsx";
+export { default as Edit } from "./Edit.tsx";
 export { default as Help } from "./Help.tsx";
+export { default as LeftArrow } from "./LeftArrow.tsx";
+export { default as Link } from "./Link.tsx";
+export { default as Plus } from "./Plus.tsx";
+export { default as Reload } from "./Reload.tsx";
 export { default as RightArrow } from "./RightArrow.tsx";
 export { default as Search } from "./Search.tsx";
+export { default as Trash } from "./Trash.tsx";
 export { default as Warn } from "./Warn.tsx";