+++ packages/wow-ui/src/components/Button/index.tsx
@@ -5,7 +5,6 @@ import { styled } from "@styled-system/jsx";
 import type { CSSProperties, ElementType, ReactNode } from "react";
 import { forwardRef } from "react";
 
-import useButton from "@/hooks/useButton";
 import type {
   PolymorphicComponentProps,
   PolymorphicComponentPropsWithRef,
@@ -16,14 +15,11 @@ import type {
  * @description 버튼 컴포넌트의 속성을 정의합니다.
  *
  * @param {ReactNode} children - 버튼의 자식 요소.
+ * @param {string} [subText] - 버튼의 하단에 위치할 보조 텍스트.
  * @param {boolean} [disabled] - 버튼이 비활성화되어 있는지 여부.
  * @param {"lg" | "sm"} [size] - 버튼의 크기.
- * @param {"solid" | "outline"} [variant] - 버튼의 종류.
- * @param {() => void} [onKeyUp] - 버튼에 포커스 된 상태에서 엔터 키 또는 스페이스 바를 뗐을 때 동작할 이벤트.
- * @param {() => void} [onKeyDown] - 버튼에 포커스 된 상태에서 엔터 키 또는 스페이스 바를 누르고 있는 동안 동작할 이벤트.
- * @param {() => void} [onMouseLeave] - 버튼의 영역에서 마우스가 벗어났을 때 동작할 이벤트.
- * @param {() => void} [onPointerDown] - 버튼에 포커스 된 상태에서 마우스 또는 터치로 누르고 있는 동안 동작할 이벤트.
- * @param {() => void} [onPointerUp] - 버튼에 포커스 된 상태에서 마우스 또는 터치를 뗐을 때 동작할 이벤트.
+ * @param {"solid" | "outline" | "sub"} [variant] - 버튼의 종류.
+ * @param {ReactNode} [icon] - 버튼의 좌측에 들어갈 아이콘.
  * @param {CSSProperties} [style] - 버튼의 커스텀 스타일.
  * @param {string} [className] - 버튼에 전달하는 커스텀 클래스.
  * @param {ComponentPropsWithoutRef<T>} rest 렌더링된 요소 또는 컴포넌트에 전달할 추가 props.
@@ -32,14 +28,11 @@ import type {
 
 export interface CustomButtonProps {
   children: ReactNode;
+  subText?: string;
   disabled?: boolean;
   size?: "lg" | "sm";
-  variant?: "solid" | "outline";
-  onKeyUp?: () => void;
-  onKeyDown?: () => void;
-  onMouseLeave?: () => void;
-  onPointerDown?: () => void;
-  onPointerUp?: () => void;
+  variant?: "solid" | "outline" | "sub";
+  icon?: ReactNode;
   style?: CSSProperties;
   className?: string;
 }
@@ -56,62 +49,35 @@ type ButtonComponent = <C extends ElementType = "button">(
 const Button: ButtonComponent & { displayName?: string } = forwardRef(
   <C extends ElementType = "button">(
     {
-      as,
+      asProp,
       children,
+      subText,
       disabled = false,
       size = "lg",
       variant = "solid",
-      onKeyUp,
-      onKeyDown,
-      onMouseLeave,
-      onPointerDown,
-      onPointerUp,
+      icon,
       ...rest
     }: ButtonProps<C>,
     ref?: PolymorphicRef<C>
   ) => {
-    const Component = as || "button";
-
-    const {
-      pressed,
-      handleKeyDown,
-      handleKeyUp,
-      handlePointerDown,
-      handlePointerUp,
-      handleMouseLeave,
-    } = useButton({
-      disabled,
-      onMouseLeave,
-      onKeyUp,
-      onKeyDown,
-      onPointerDown,
-      onPointerUp,
-    });
+    const Component = asProp || "button";
 
     return (
       <Component
         aria-disabled={disabled}
-        aria-pressed={pressed}
         disabled={disabled}
         ref={ref}
         className={ButtonStyle({
-          size,
+          size: variant === "sub" ? "sm" : size,
           variant,
         })}
-        onKeyDown={handleKeyDown}
-        onKeyUp={handleKeyUp}
-        onMouseLeave={handleMouseLeave}
-        onPointerDown={handlePointerDown}
-        onPointerUp={handlePointerUp}
         {...rest}
       >
-        <styled.span
-          {...(typeof children === "string" && {
-            textStyle: size === "lg" ? "label1" : "label2",
-          })}
-        >
+        <styled.span className={ContentStyle({ size })}>
+          {icon}
           {children}
         </styled.span>
+        {subText && <styled.span textStyle="label3">{subText}</styled.span>}
       </Component>
     );
   }
@@ -120,8 +86,10 @@ const Button: ButtonComponent & { displayName?: string } = forwardRef(
 const ButtonStyle = cva({
   base: {
     display: "flex",
+    flexDirection: "column",
     alignItems: "center",
     justifyContent: "center",
+    gap: "xs",
 
     cursor: "pointer",
   },
@@ -130,7 +98,6 @@ const ButtonStyle = cva({
       lg: {
         width: "100%",
         maxWidth: { lgOnly: 316 },
-        height: "3rem",
         padding: "1rem",
         borderRadius: "md",
       },
@@ -145,13 +112,14 @@ const ButtonStyle = cva({
         color: "textWhite",
 
         _disabled: {
-          background: "darkDisabled",
-          cursor: "not-allowed",
+          background: "monoBackgroundPressed",
+          color: "outline",
+          pointerEvents: "none",
         },
         _hover: {
           shadow: "blue",
         },
-        _pressed: {
+        _active: {
           background: "bluePressed",
         },
       },
@@ -166,16 +134,31 @@ const ButtonStyle = cva({
         _disabled: {
           borderColor: "darkDisabled",
           color: "darkDisabled",
-          cursor: "not-allowed",
+          pointerEvents: "none",
         },
         _hover: {
           borderColor: "blueHover",
           color: "blueHover",
         },
-        _pressed: {
+        _active: {
           borderColor: "bluePressed",
           background: "blueBackgroundPressed",
           color: "bluePressed",
+        },
+      },
+      sub: {
+        background: "blueBackgroundPressed",
+        color: "primary",
+
+        _disabled: {
+          color: "blueDisabled",
+          pointerEvents: "none",
+        },
+        _hover: {
+          shadow: "blue",
+        },
+        _active: {
+          background: "blueDisabled",
         },
       },
     },
@@ -192,7 +175,7 @@ const ButtonStyle = cva({
           borderColor: "textBlack",
           color: "textBlack",
         },
-        _pressed: {
+        _active: {
           borderColor: "outline",
           background: "monoBackgroundPressed",
           color: "textBlack",
@@ -200,6 +183,26 @@ const ButtonStyle = cva({
       },
     },
   ],
+});
+
+const ContentStyle = cva({
+  base: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  variants: {
+    size: {
+      lg: {
+        gap: "xs",
+        textStyle: "label1",
+      },
+      sm: {
+        gap: "xxs",
+        textStyle: "label2",
+      },
+    },
+  },
 });
 
 Button.displayName = "Button";