+++ packages/wow-ui/src/components/DropDown/DropDownOption.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import { cva } from "@styled-system/css";
+import { cva, cx } from "@styled-system/css";
 import { styled } from "@styled-system/jsx";
-import type { ReactNode } from "react";
+import type { CSSProperties, ReactNode } from "react";
 import { forwardRef, useCallback, useEffect } from "react";
 
 import { useDropDownContext } from "@/components/DropDown/context/DropDownContext";
@@ -14,17 +14,22 @@ import { useCollection } from "./context/CollectionContext";
  *
  * @param {string} value 옵션의 값입니다.
  * @param {() => void} [onClick] 옵션이 클릭되었을 때 호출되는 함수입니다.
- * @param {React.ReactNode} [text] 드롭다운 옵션에 들어갈 텍스트.
+ * @param {React.ReactNode} text 드롭다운 옵션에 들어갈 텍스트.
+ * @param {CSSProperties} [style] 드롭다운 옵션의 커스텀 스타일.
+ * @param {string} [className] 드롭다운 옵션에 전달하는 커스텀 클래스.
  */
 export interface DropDownOptionProps {
   value: string;
   onClick?: () => void;
   text: ReactNode;
+  style?: CSSProperties;
+  className?: string;
 }
 
 const DropDownOption = forwardRef<HTMLLIElement, DropDownOptionProps>(
-  function Option({ value, onClick, text }, ref) {
-    const { focusedValue, selectedValue, handleSelect } = useDropDownContext();
+  function Option({ value, onClick, text, className, ...rest }, ref) {
+    const { focusedValue, selectedValue, handleSelect, dropdownId } =
+      useDropDownContext();
     const isSelected = selectedValue === value;
     const isFocused = focusedValue !== null && focusedValue === value;
 
@@ -47,16 +52,20 @@ const DropDownOption = forwardRef<HTMLLIElement, DropDownOptionProps>(
 
     return (
       <styled.li
-        id={`dropdown-option-${value}`}
+        id={`${dropdownId}-option-${value}`}
         ref={ref}
         role="option"
         tabIndex={isSelected ? 0 : -1}
-        className={optionStyle({
-          type: isSelected ? "selected" : isFocused ? "focused" : "default",
-        })}
+        className={cx(
+          optionStyle({
+            type: isSelected ? "selected" : isFocused ? "focused" : "default",
+          }),
+          className
+        )}
         onClick={() => {
           handleOptionClick(value, onClick);
         }}
+        {...rest}
       >
         {text}
       </styled.li>