+++ packages/wow-ui/src/components/DropDown/index.tsx
@@ -90,25 +90,20 @@ const DropDown = ({
   onChange,
   ...rest
 }: DropDownProps) => {
+  const defaultId = useId();
+  const dropdownId = id ?? `dropdown-${defaultId}`;
+
   const dropdownState = useDropDownState({
     value,
     defaultValue,
     onChange,
   });
 
-  const defaultId = useId();
-  const dropdownId = id ?? `dropdown-${defaultId}`;
-
   return (
-    <DropDownContext.Provider value={dropdownState}>
+    <DropDownContext.Provider value={{ ...dropdownState, dropdownId }}>
       <CollectionProvider>
-        <DropDownWrapper
-          dropdownId={dropdownId}
-          hasCustomTrigger={!!trigger}
-          {...rest}
-        >
+        <DropDownWrapper hasCustomTrigger={!!trigger} {...rest}>
           <DropDownTrigger
-            dropdownId={dropdownId}
             label={label}
             placeholder={placeholder}
             trigger={trigger}