+++ packages/wow-theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # wowds-theme
 
+## 0.1.3
+
+### Patch Changes
+
+- f880aa6: Tag 컴포넌트를 구현합니다.
+
 ## 0.1.2
 
 ### Patch Changes