+++ packages/wow-tokens/CHANGELOG.md
@@ -1,5 +1,17 @@
 # wowds-tokens
 
+## 0.1.3
+
+### Patch Changes
+
+- 1632731: Button 컴포넌트의 추가 스펙을 구현합니다.
+
+## 0.1.2
+
+### Patch Changes
+
+- f880aa6: Tag 컴포넌트를 구현합니다.
+
 ## 0.1.1
 
 ### Patch Changes